rename files
https://coveord.atlassian.net/browse/KIT-441
